Refactor ropes
This moves the two rope variants into a ropes module. The old import
paths have been maintained by re-exporting.

This patch also extracts the common methods from the Rope impls to a
macro. This lets us write shared methods once, rather than have to
duplicate for both variants.
